// src/app/models/fty.ts
export class FTY {
    id?: string;
    fty?: string = '';
    vendId?: string = '';
    ftyDesc?: string = '';
    ftyGroup?: string = '';
    ftyStatus?: boolean = true;
    portOfLoading?: string = ''
    telexReleaseTerm?: string = '';
    nyTerms?: string = '';
    caTerms?: string = '';
    lclSurcharge?: string = ''
    termNotes?: string = ''
}